python-sdk: inline turn handles into api facade
